requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argilla==1.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
effdet==0.3.0
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.3
fastapi==0.95.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.136
layoutparser==0.3.4
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.22.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.0
pypandoc==1.11
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
timm==0.6.13
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.11
unstructured-inference==0.3.2
urllib3==1.26.15
uvicorn==0.21.1
Wand==0.6.11
wcwidth==0.2.6
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
